token_platform/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
channels==4.1.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
daphne==4.1.2
cryptography==42.0.7
factory_boy==3.3.0
pytest-cov==5.0.0
sentry-sdk==2.9.0
